// KairosFrontend/src/app/services/model/page.ts
/**
 * Kairos Backend API
 * This API provides endpoints for the Kairos Backend application.
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { PageableObject } from './pageableObject';
import { SortObject } from './sortObject';

export interface Page { 
    totalPages?: any;
    totalElements?: any;
    last?: any;
    pageable?: PageableObject;
    first?: any;
    numberOfElements?: any;
    size?: any;
    content?: any;
    number?: any;
    sort?: SortObject;
    empty?: any;
}